feat: splitting layouts
