+++ apps/storefront/app/components/Hero.tsx
@@ -7,7 +7,7 @@ import type {
   Video as MediaVideo,
 } from '@shopify/hydrogen/storefront-api-types';
 import {Heading, Text, Link} from '~/components';
-import type {CollectionHero} from '~/routes/($lang)/index';
+import type {CollectionHero} from '~/routes/__store/($lang)/index';
 
 /**
  * Hero component that renders metafields attached to collection resources
