+++ apps/storefront/app/components/SortFilter.tsx
@@ -20,7 +20,7 @@ import type {
 import {
   AppliedFilter,
   SortParam,
-} from '~/routes/($lang)/collections/$collectionHandle';
+} from '~/routes/__store/($lang)/collections/$collectionHandle';
 
 type Props = {
   filters: Filter[];
