+++ apps/storefront/app/routes/__store/($lang)/account/__private/logout.ts
