+++ apps/storefront/app/routes/__store/($lang)/account/__public/activate.$id.$activationToken.tsx
