+++ apps/storefront/app/routes/__store/($lang)/account/__public/reset.$id.$resetToken.tsx
