+++ apps/storefront/app/routes/__store/($lang)/api/products.tsx
