+++ apps/storefront/app/routes/__store/($lang)/collections/$collectionHandle.tsx
