+++ apps/storefront/app/routes/__store/($lang)/collections/all.tsx
