+++ apps/storefront/app/routes/__store/($lang)/discount.$code.tsx
