+++ apps/storefront/app/routes/__store/($lang)/journal/index.tsx
