+++ apps/storefront/app/routes/__store/($lang)/policies/$policyHandle.tsx
