+++ apps/storefront/app/routes/__store/($lang)/policies/index.tsx
