--- packages/vite-plugin-validate-environment/.gitignore
@@ -1,6 +1,0 @@
-# dependencies
-node_modules
-.pnp
-.pnp.js
-
-lib
