--- packages/vite-plugin-validate-environment/package.config.ts
@@ -1,12 +1,0 @@
-// package.config.ts
-
-import {defineConfig} from '@sanity/pkg-utils'
-
-export default defineConfig({
-  minify: true,
-  dist: './lib',
-  runtime: 'node',
-  extract: { rules: { 
-    'ae-missing-release-tag': 'warn',
-  } }
-})
