--- packages/vite-plugin-validate-environment/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "vite-plugin-validate-environment",
-  "version": "",
-  "type": "commonjs",
-  "exports": {
-    ".": {
-      "types": "./lib/index.d.ts",
-      "source": "./src/index.ts",
-      "require": "./lib/index.js",
-      "default": "./lib/index.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "main": "./lib/index.js",
-  "source": "./src/index.ts",
-  "types": "./lib/index.d.ts",
-  "scripts": {
-    "build": "pkg-utils build"
-  },
-  "devDependencies": {
-    "@sanity/pkg-utils": "^2.2.1",
-    "typescript": "^4.7.2",
-    "vite": "^2.9.13"
-  },
-  "peerDependencies": {
-    "vite": "^2"
-  },
-  "license": ""
-}
