--- packages/vite-plugin-validate-environment/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./lib",
-    "target": "es2020",
-    "lib": [
-      "es2020", "DOM"
-    ],
-    "types": ["node"],
-    "module": "ES2020",
-    "moduleResolution": "Node",
-    "strict": true,
-    "sourceMap": true,
-    "declaration": true,
-  },
-  "exclude": [
-    "node_modules",
-    "lib"
-  ],
-  "include": [
-    "src/**/*.ts"
-  ]
-}
